Refactor: chagne rename and package path of Tag class
