+++ src/main/java/com/gdscsolutionchallenge/shareBite/postTag/entity/PostTag.java
@@ -1,7 +1,7 @@
-package com.gdscsolutionchallenge.shareBite.postFoodCategory;
+package com.gdscsolutionchallenge.shareBite.postTag.entity;
 
-import com.gdscsolutionchallenge.shareBite.foodCategory.FoodCategory;
-import com.gdscsolutionchallenge.shareBite.post.Post;
+import com.gdscsolutionchallenge.shareBite.tag.entity.Tag;
+import com.gdscsolutionchallenge.shareBite.post.entity.Post;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -10,8 +10,8 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-@Entity
-public class PostFoodCategory {
+@Entity(name = "POST_TAGS")
+public class PostTag {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long postFoodCategoryId;
@@ -22,5 +22,5 @@ public class PostFoodCategory {
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "FOOD_CATEGORY_ID")
-    private FoodCategory foodCategory;
+    private Tag tag;
 }
